# Conflicts:
#	.vscode/launch.json
#	meshtastic/protobuf/config_pb2.py
#	meshtastic/protobuf/mesh_pb2.py
#	meshtastic/protobuf/mesh_pb2.pyi

+++ bin/regen-protobufs.sh
@@ -15,7 +15,7 @@ echo "Fixing up protobuf paths in ${TMPDIR} temp directory"
 
 
 # Ensure a clean build
-rm -r "${TMPDIR}"
+[ -e "${TMPDIR}" ] && rm -r "${TMPDIR}"
 
 INDIR=${TMPDIR}/in/meshtastic/protobuf
 OUTDIR=${TMPDIR}/out
